Refactors modal
Refactors modal component, changing its content depending on state (select date, confirm reservation, show results).
